src/requirements.txt
numpy
torch
torchvision
pyyaml
opencv-python
pillow
tqdm